Curry throws an error if passed more than one argument
As per conversations with @guipn, formerly the API allowed calls to be made
like so:

```javascript
    var sum = function(){

        var args = [].slice.call(arguments);

        return args.reduce(function(a, b){ return a + b }, 0);

    };

    var sum5 = sum.curry(5, 2, 3);

```

The first argument would be considered the depth, and the subsequent ones
as arguments (leaving 3 arguments left until the original function is applied.

This was considered confusing, so now the above throws an error.